requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
google-generativeai==0.8.3
pydantic==2.5.2
python-multipart==0.0.6
Pillow==10.1.0
PyMuPDF==1.23.7
python-dotenv==1.0.0
